+++ packages/demo/__stories__/plugin-video.tsx
@@ -3,9 +3,17 @@ import * as React from 'react'
 
 import { EditorStory } from '../src'
 
-storiesOf('Plugins/Video', module).add('Initial State', () => {
-  const state = {
-    plugin: 'video'
-  }
-  return <EditorStory initialState={state} />
-})
+storiesOf('Plugins/Video', module)
+  .add('Initial State', () => {
+    const state = {
+      plugin: 'video'
+    }
+    return <EditorStory initialState={state} />
+  })
+  .add('Prefilled', () => {
+    const state = {
+      plugin: 'video',
+      state: 'https://www.youtube.com/watch?v=SCJ7nzKwnYo'
+    }
+    return <EditorStory initialState={state} />
+  })