+++ packages/plugin-rows/src/renderer.tsx
@@ -1,11 +1,8 @@
 import { StatefulPluginEditorProps } from '@edtr-io/core'
 import { rowsState } from '@edtr-io/plugin-rows'
 import * as React from 'react'
-import { styled } from '@edtr-io/ui'
 
-export const Row = styled.div({
-  margin: '5px 0'
-})
+import { RowContainer } from './row'
 
 export const RowsRenderer = (
   props: StatefulPluginEditorProps<typeof rowsState>
@@ -13,7 +10,7 @@ export const RowsRenderer = (
   return (
     <React.Fragment>
       {props.state().map(row => {
-        return <Row key={row.id}>{row.render()}</Row>
+        return <RowContainer key={row.id}>{row.render()}</RowContainer>
       })}
     </React.Fragment>
   )