+++ packages/plugin-text/src/plugins/link.tsx
@@ -57,12 +57,22 @@ export interface LinkPluginOptions {
 const DefaultEditorComponent: React.FunctionComponent<
   NodeEditorProps
 > = props => {
-  const { attributes, children, node } = props
+  const { attributes, children, node, isSelected } = props
   const inline = node
   const href = inline.data.get('href')
 
   return (
-    <a {...attributes} href={href}>
+    <a
+      {...attributes}
+      href={href}
+      style={
+        isSelected
+          ? {
+              textDecoration: 'underline'
+            }
+          : undefined
+      }
+    >
       {children}
     </a>
   )
@@ -100,7 +110,7 @@ const DefaultControlsComponent: React.FunctionComponent<
   return (
     <React.Fragment>
       {props.children}
-      {isLink(editor) && !overlayContext.visible ? (
+      {!props.readOnly && isLink(editor) && !overlayContext.visible ? (
         <InlineOverlay
           key={`inlineoverlay${inline.key}`}
           onEdit={overlayContext.show}
@@ -111,7 +121,7 @@ const DefaultControlsComponent: React.FunctionComponent<
           </a>
         </InlineOverlay>
       ) : null}
-      {isLink(editor) ? (
+      {!props.readOnly && isLink(editor) ? (
         <Overlay key={`overlay${inline.key}`} onClose={() => editor.focus()}>
           <AutoFocusInput
             label="URL"