+++ packages/plugin-video/src/editor.tsx
@@ -1,5 +1,5 @@
-import { OverlayContext, StatefulPluginEditorProps } from '@edtr-io/core'
-import { Overlay, AutoFocusInput } from '@edtr-io/ui'
+import { StatefulPluginEditorProps } from '@edtr-io/core'
+import { Overlay, AutoFocusInput, ContainerWithConfigButton } from '@edtr-io/ui'
 import * as React from 'react'
 
 import { VideoRenderer } from './renderer'
@@ -10,12 +10,11 @@ export const VideoEditor = ({
   focused,
   state
 }: StatefulPluginEditorProps<typeof videoState>) => {
-  const overlayContext = React.useContext(OverlayContext)
   return (
     <React.Fragment>
-      <div onClick={editable ? overlayContext.show : undefined}>
+      <ContainerWithConfigButton>
         <VideoRenderer state={state} disableCursorEvents={editable} />
-      </div>
+      </ContainerWithConfigButton>
       {focused ? (
         <Overlay>
           <AutoFocusInput