+++ packages/plugin-video/src/index.tsx
@@ -1,10 +1,14 @@
 import { StatefulPlugin, StateType } from '@edtr-io/core'
+import * as React from 'react'
 
 import { VideoEditor } from './editor'
+import { VideoRenderer } from './renderer'
 
 export const videoState = StateType.string()
 export const videoPlugin: StatefulPlugin<typeof videoState> = {
-  Component: VideoEditor,
+  //eslint-disable-next-line react/display-name
+  Component: props =>
+    props.editable ? <VideoEditor {...props} /> : <VideoRenderer {...props} />,
   state: videoState,
   onPaste(clipboardData: DataTransfer) {
     const value = clipboardData.getData('text')